moved login route to own file, moved encrypt password to helpers folder
