+++ routes/index.js
@@ -12,6 +12,7 @@ const deleteChild = require("./delete-child");
 const getChildTransactions = require('./get-child-transactions');
 const confirmParent = require("./confirm-parent");
 const rejectParent = require("./reject-parent");
+const login = require('./login');
 
 const routesList = [];
 
@@ -28,7 +29,8 @@ routesList.push(deleteChild);
 routesList.push(getChildTransactions);
 routesList.push(confirmParent);
 routesList.push(rejectParent);
-routesList.push(getUserTransactions)
+routesList.push(getUserTransactions);
+routesList.push(login);
 
 function useCustomRoutes(app, db) {
     // tell express server to use routes
